requirements.txt
# 核心依赖
pydantic>=2.0.0,<2.8
pyyaml>=6.0
click>=8.0.0
loguru>=0.7.0

# 异步支持
aiohttp>=3.8.0

# 本地向量数据库 - 使用较稳定的版本
chromadb>=0.4.0,<0.5.0

# 数据处理
numpy>=1.24.0,<2.0
pandas>=2.0.0

# 报告生成
jinja2>=3.1.0
plotly>=5.15.0

# 测试
pytest>=7.0.0
pytest-asyncio>=0.21.0
